+++ lib/magic/decoratable.rb
@@ -3,12 +3,13 @@
 module Magic
 	module Decoratable
 		def decorate   = decorator&.new self
-		def decorate!  = decorate || raise(Lookup::Error.for self, Decorator)
+		def decorate!  = decorate || raise(Lookup::Error.for self, decorator_base)
 		def decorated  = decorate || self
 		def decorated? = false
 
 		private
 
-		def decorator = Decorator.for self.class
+		def decorator      = decorator_base.for self.class
+		def decorator_base = Decorator
 	end
 end